New module injection API
